+++ R/pattern-array-magick.R
@@ -12,7 +12,7 @@
 #' @param fill Fill colour
 #' @return A grid grob object invisibly.  If `draw` is `TRUE` then also draws to the graphic device as a side effect.
 #' @examples
-#'   if (require("magick")) {
+#'   if (requireNamespace("magick")) {
 #'     x_hex <- 0.5 + 0.5 * cos(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
 #'     y_hex <- 0.5 + 0.5 * sin(seq(2 * pi / 4, by = 2 * pi / 6, length.out = 6))
 #'     grid.pattern_magick(x_hex, y_hex, type="octagons", fill="blue", scale=2)